Add API provider
The API provider makes use of the cacti API to manage cati hosts. This
way puppet does not necessarily need to know the database backend of
cacti or the database password.

Because the API is written in PHP the helperscript apihelper.php needs
to do the necessary translations between PHP world and ruby world.

This commit is just a raw sketch and the provider is far from beeing
functional. The only bit that should already work is prefetching so

    puppet resource cacti_host

Should already show results.