Write JSON to file instead of stdout
When executing the apihelper we may want to print some debug messages or
php itself may print some warnings (mostly deprecation warnings if you
are using an older cacti version) on stdout. To make the communcation
between the provider and the apihelper a litte bit more reliable, puppet
will now create a tempfile and asks the apihelper to write the JSON output
to that tempfile which can the be processed in puppet again.

This way the apihelper is able to print normal messages on the console
whitout interfering with the JSON output.