Remove default values form cacti_host type
Using default values for all properties makes sense when creating a new
cacti_host but it also forces the user to manage all properties through
puppet. So if someone adds a note to a device through the cacti web
application and you are also managing this device through puppet, puppet
will reset the note even if you have not specified one explicitly
in your manifest because the default value is an empty string.

The default values are now handled by the api provider so they will only
be set if there are no current values (which is only the case when the
resource is not yet present). This way you can even specify

    cacti_host { "puppet":
      ensure   => present,
      hostname => "puppet.example.com"
    }

And puppet will make sure the resource is present. If the resource is
absent it will be created with default values for snmp version,
community string etc, but if the resource is already present, puppet
will not touch any of these attributes (except the properties you have
explicitly specified of course (only hostname in the example above)).