Add warning if apihelper does not retrun json
In case the mysql server is not running the require statements inside
the apihelper.php script do not work. Unforunately the script does
return with a zero exit code so we can only recognize the failure by
looking at the outputfile because it will be empty and will not contain
valid json data.

Rescue the json parse error and provide a bit more details so the user
can hopefully understand what is going on.