Add the ability to pretty print json output
The apihelper.php script can also be used as a normal user to query
hosts without using the web application. Using apihelper.php does also
make sense to debug problems. In this cases it is a lot easier to
understand the JSON output if it is spread on multiple lines with proper
indention.