Update property hash after setting values
The cacti_plugin provider and the cacti_setting provider do use
prefetching so the current value is stored in @property_hash after the
prefetch phase.

If we later updated a value because a property was out of sync puppet
would not update the property_hash so if the property was queried later
we would still return the old value. This is fixed now.